% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher's Method for Aggregating p-values}
\usage{
fisher(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated using Fisher's Method. NAs and values outside the interval [0,1] will be removed.}
}
\value{
A single scalar aggregated p-value (between 0 and 1).
}
\description{
Aggregate a given set of p-values with equal weights. This is equivalent to using Lancaster's Method with all p-values weighted by 2.
}
\examples{
fisher(c(.1, .2, .3))

}
\references{
Fisher, R. A. (1992). Statistical methods for research workers. In \emph{Breakthroughs in statistics} (pp. 66-70). Springer, New York, NY.
}
\seealso{
\code{\link{lancaster}}, \code{\link{stouffer}}, \code{\link{sidak}}
}
\author{
Rupam Bhattacharyya, \email{rupamb@umich.edu}
}
