% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupamb625.R
\docType{package}
\name{rupamb625}
\alias{rupamb625}
\title{rupamb625: A package for aggregating p-values}
\description{
Aggregate a given set of p-values using one of the following four methods:
Fisher's, Lancaster's, Stouffer's or Sidak's.
}
\section{Functions in rupamb625}{

\code{\link{fisher}}, \code{\link{lancaster}}, \code{\link{stouffer}}, \code{\link{sidak}}
}

\author{
Rupam Bhattacharyya, \email{rupamb@umich.edu}
}
