% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidak.R
\name{sidak}
\alias{sidak}
\title{Sidak's Method for Aggregating p-values}
\usage{
sidak(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated using Sidak's Method. NAs and values outside the interval [0,1] will be removed.}
}
\value{
A single scalar aggregated p-value (between 0 and 1).
}
\description{
Aggregate a set of p-values by adjusting the minimum p-value w.r.t. the number of p-values being adjusted.
}
\examples{
sidak(c(.1, .2, .3))

}
\references{
Šidák, Z. (1967). Rectangular confidence regions for the means of multivariate normal distributions. \emph{Journal of the American Statistical Association}, 62(318), 626-633.
}
\seealso{
\code{\link{lancaster}}, \code{\link{stouffer}}, \code{\link{fisher}}
}
\author{
Rupam Bhattacharyya, \email{rupamb@umich.edu}
}
