% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stouffer.R
\name{stouffer}
\alias{stouffer}
\title{Stouffer's Method for Aggregating p-values}
\usage{
stouffer(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated using Stouffer's Method. NAs and values outside the interval [0,1] will be removed.}
}
\description{
Aggregate a set of one-sided p-values by using standard normal transformations.
}
\examples{
stouffer(c(.05, .2, .5))
}
