% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher's Method for Aggregating p-values}
\usage{
fisher(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated using Fisher's Method. NAs and values outside the interval [0,1] will be removed.}
}
\description{
Aggregate a given set of p-values with equal weights. This is equivalent to using Lancaster's Method with all p-values weighted by 2.
}
\examples{
fisher(c(.1, .2, .3))
}
