% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidak.R
\name{sidak}
\alias{sidak}
\title{Sidak's Method for Aggregating p-values}
\usage{
sidak(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated using Sidak's Method. NAs and values outside the interval [0,1] will be removed.}
}
\description{
Aggregate a set of p-values by adjusting the minimum p-value w.r.t. the number of p-values being adjusted.
}
\examples{
sidak(c(.1, .2, .3))
}
