% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stouffer.R
\name{stouffer}
\alias{stouffer}
\title{Stouffer's Method for Aggregating p-values}
\usage{
stouffer(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated using Stouffer's Method. NAs and values outside the interval [0,1] will be removed.}
}
\value{
A single scalar aggregated p-value (between 0 and 1).
}
\description{
Aggregate a set of one-sided p-values by using standard normal transformations.
}
\examples{
stouffer(c(.05, .2, .5))

}
\references{
Stouffer, S. A., Suchman, E. A., DeVinney, L. C., Star, S. A., & Williams Jr, R. M. (1949). The american soldier: Adjustment during army life.(studies in social psychology in world war ii), vol. 1.
}
\seealso{
\code{\link{lancaster}}, \code{\link{fisher}}, \code{\link{sidak}}
}
\author{
Rupam Bhattacharyya, \email{rupamb@umich.edu}
}
